requirements.txt
pyyaml
jinja2
weasyprint
flask
cryptography
anthropic
openai
google-genai>=1.0.0
language-tool-python
pdfplumber
daytona-sdk
psycopg2-binary
gunicorn
mcp[cli]
uvicorn
requests
beautifulsoup4
